dia receta y alimento

+++ src/main/java/com/jkstic/testHw/services/impl/AlimentosServiceImpl.java
@@ -1,6 +1,9 @@
 package com.jkstic.testHw.services.impl;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +11,9 @@ import org.springframework.stereotype.Service;
 
 import com.jkstic.testHw.dto.AlimentoDto;
 import com.jkstic.testHw.models.Alimentos;
+import com.jkstic.testHw.models.Stock;
 import com.jkstic.testHw.repositories.AlimentoRepository;
+import com.jkstic.testHw.repositories.StockRepository;
 import com.jkstic.testHw.services.AlimentosService;
 import com.jkstic.testHw.transformer.AlimentosMapper;
 
@@ -16,20 +21,34 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
-public class AlimentosServiceImpl implements AlimentosService{
-	
+public class AlimentosServiceImpl implements AlimentosService {
+
 	@Autowired
 	private AlimentosMapper alimentosMapper;
-	
+
 	@Autowired
 	private AlimentoRepository alimentoRepository;
-	
-	
+
+	@Autowired
+	private StockRepository stockRepository;
 
 	@Override
-	public AlimentoDto createAliment(AlimentoDto aliment){
-		log.info("createAliment: ", aliment.getNombre());
-		return alimentosMapper.alimentosToAlimentoDto(alimentoRepository.save(alimentosMapper.alimentoDtotoAlimentos(aliment)));
+	public AlimentoDto createAliment(AlimentoDto aliment) throws Exception {
+		log.info("createAliment: ", aliment.getNombre().toString());
+		Alimentos alimentoTarget = alimentosMapper.alimentoDtotoAlimentos(aliment);
+
+		try {
+
+			Stock stockTarget = stockRepository
+					.save(Stock.builder().cantidadStock(aliment.getStock().getCantidadStock()).build());
+
+			alimentoTarget.setStock(stockTarget);
+
+			Alimentos alimentoSaved = alimentoRepository.save(alimentoTarget);
+			return alimentosMapper.alimentosToAlimentoDto(alimentoSaved);
+		} catch (Exception e) {
+			throw new Exception(e.getMessage());
+		}
 	}
 
 	@Override
@@ -41,7 +60,7 @@ public class AlimentosServiceImpl implements AlimentosService{
 	@Override
 	public AlimentoDto verAliment(String nombre) {
 		log.info("verAliment: ", nombre);
-		return alimentosMapper.alimentosToAlimentoDto(alimentoRepository.findByNombre(nombre));
+		return alimentosMapper.alimentosToAlimentoDto(alimentoRepository.findByNombre(nombre).get());
 	}
 
 	@Override
@@ -51,13 +70,33 @@ public class AlimentosServiceImpl implements AlimentosService{
 		alimentToModificated.setNombre(aliment.getNombre());
 		alimentToModificated.setCantidad(aliment.getCantidad());
 		alimentToModificated.setMedida(aliment.getMedida());
+
+		if (alimentToModificated.getStock() == null) {
+			Stock stockTarget = stockRepository
+					.save(Stock.builder().cantidadStock(aliment.getStock().getCantidadStock()).build());
+			alimentToModificated.setStock(stockTarget);
+		}else {
+			alimentToModificated.getStock().setCantidadStock(aliment.getStock().getCantidadStock());
+		}
+
 		return alimentosMapper.alimentosToAlimentoDto(alimentoRepository.save(alimentToModificated));
 	}
 
 	@Override
-	public void deleteAliment(Long id) {
+	public void deleteAliment(Long id) throws Exception {
 		log.info("deleteAliment: ", id);
-		alimentoRepository.deleteById(id);
+		Optional<Alimentos> alimento = alimentoRepository.findById(id);
+		if (alimento.isPresent()) {
+			
+			
+			alimentoRepository.deleteById(id);
+			if(alimento.get().getStock() != null) {
+				stockRepository.deleteById(alimento.get().getStock().getId());
+			}
+		}else {
+			throw new Exception("No encontrado el alimento");
+		}
+			
 	}
 
 	@Override
@@ -65,6 +104,15 @@ public class AlimentosServiceImpl implements AlimentosService{
 		log.info("verAlimentos");
 		List<Alimentos> listado = alimentoRepository.findAll();
 		return listado.stream().map(x -> alimentosMapper.alimentosToAlimentoDto(x)).collect(Collectors.toList());
+	}
+
+	@Override
+	public Map<Long, String> verIdAlimentos() {
+		Map<Long, String> result = new HashMap<Long, String>();
+		alimentoRepository.findAll().forEach(alimento -> {
+			result.put(alimento.getId(), alimento.getNombre());
+		});
+		return result;
 	}
 
 }
